requirements.txt
flask
flask-cors
numpy
opencv-python
pillow
requests
pyyaml
onnxruntime
insightface
